backend_model/requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0

# Computer Vision and ML
torch>=1.13.0
torchvision>=0.14.0
detectron2 @ git+https://github.com/facebookresearch/detectron2.git
Pillow==10.1.0
numpy==1.24.3
opencv-python==4.8.1.78

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1

# Development (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
